requirements.txt
asgiref==3.4.1
Django==3.2.18
djangorestframework==3.14.0
pika==1.3.1
python-dotenv==0.20.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.1.1
